Add 2 new GSIs for deployment (#28)
The query works for Author Filter in Guestbook after the last PR :D

I got this error, and I think it's because I don't have these 2 new GSIs
in main branch for deployment

![image](https://github.com/CodeDayLabs311/cl311/assets/93412982/a175146a-7e0a-48b3-aee9-28cf45f9668c)

I don't know if I can deploy locally by myself (I'm afraid to try, I'm
not sure if this will affect the billing :( ). But if it's safe to
deploy locally, then you can ignore this PR. Please let me know if it's
safe to deploy locally.