Editing page backend (#29)
**Editing Page Backend (Only Backend)**

Editing Page and Creating Page are using the same UI, and Zhanping is in
charge of it.

So, Sophie's task is just to implement updating data when changes happen
in the editing report form.

Implemented putReport function in IReportClient.ts, the interface for
report clients. In ReportApiClient.ts, built put endpoints to interact
with report API, etc.

I was wondering if we should have 'Date and Time Last Edited'. I decided
to build one, so I added 'dateTimeLastEdited: string;' in the
IReport.ts, our database schema.