S3 Upload Demo (#45)
I need some feedback on this. I am quite confused on how to use "bucket
and CORS". I tried to use all 6 existing s3 buckets we have right now by
adding it to the BUCKET_NAME variable in .env.local. That is the way
that the tutorial did it.

I was able to clone the vercel tutorial ,and create the S3 bucket with
approiate CORS manually on my personal AWS account. The vercel code
would work on my AWS console and local machine properly.

I think I am missing the piece in the puzzle of, how and where is the
bucket/CORS created, and what bucket name, we have throughout the
"running" phase of our codebase?

---------

Co-authored-by: Andrey Butenko <andreybutenko@users.noreply.github.com>